+++ .eslintrc.json
@@ -23,7 +23,8 @@
     "symbol-description": "off",
     "@typescript-eslint/ban-types": "off",
     "no-param-reassign": "off",
-    "no-unused-vars": "off"
+    "no-unused-vars": "off",
+    "import/prefer-default-export": "off"
   },
   "overrides": [{
     "files": ["__tests__/**/*"],
