+++ __tests__/10_basic_spec.ts
@@ -1,4 +1,4 @@
-import memoize from '../src/index';
+import { memoize } from '../src/index';
 
 describe('basic spec', () => {
   it('x.a', () => {
