--- src/wrapper.cjs
@@ -1,4 +1,0 @@
-const { default: defaultFn, ...rest } = require('./index.umd.js');
-
-module.exports = defaultFn;
-Object.assign(module.exports, rest);
